-Restructured Components -Added SVG -Fixed height helper function -Added Conditional Render for Alert -README -Started to fix Card component/styling
ToDo:
-checkGame to push grave into hand and process wins (will do in morning)
