+++ client/src/components/Play/Deck/deck.css
@@ -1,0 +1,57 @@
+/* Deck Cards */
+.player1Deck, .player2Deck{
+  width: 150px !important;
+  height: 250px !important;
+  /* background-image: url(./images/logoNBA.png); */
+  background-color: black;
+  background-size: contain;
+  background-position: center;
+  border: 1px solid;
+  position: absolute;
+  border-radius: 10px;
+  z-index: 2;
+}
+
+.active {
+  border: 5px solid yellow
+}
+
+.active:hover {
+  cursor: pointer;
+}
+
+.player1Deck{
+  bottom: 10%;
+  left: 8%;
+}
+
+.player2Deck{
+  bottom: 10%;
+  right: 8%;
+}
+
+.card-img-top{
+  height: 200px;
+  width: 100px;
+}
+
+/* Team Colors */
+.Lakers {
+  /* background: linear-gradient(gold, purple); */
+  background-image: url("../../../images/Lakers.png");
+}
+
+.Raptors {
+  /* background: linear-gradient(#CE1141,#000000,#A1A1A4); */
+  background-image: url("../../../images/Raptors.png");
+}
+
+.Heat {
+  /* background: linear-gradient(#98002E, #F9A01B, #000000) ; */
+  background-image: url("../../../images/Heat.png");
+}
+
+.Rockets {
+  /* background: linear-gradient(#C4CED4, #CE1141, #000000) */
+  background-image: url("../../../images/Rockets.png");
+}
