--- client/src/components/Play/Graveyard.js
@@ -1,10 +1,0 @@
-import React from "react";
-import "./graveyard.css";
-
-function Graveyard({ className }) {
-  return (
-    <div className={`${className}Grave`} />
-  );
-}
-
-export default Graveyard;
