+++ client/src/components/Play/Graveyard/Graveyard.js
@@ -1,0 +1,10 @@
+import React from "react";
+import "./graveyard.css";
+
+function Graveyard({ className }) {
+  return (
+    <div className={`${className}Grave`} />
+  );
+}
+
+export default Graveyard;
