--- client/src/components/Play/card.scss
@@ -1,254 +1,0 @@
-$lakers-pri: #FDB927;
-$lakers-sec: #552583;
-$heat-pri: #98002E;
-$heat-sec: #000000;
-$raptors-pri: #000000;
-$raptors-sec: #CE1141;
-$rockets-pri: #CE1141;
-$rockets-sec: #C4CED4;
-
-/* Hand Cards */
-.player1Card, .player2Card{
-    position: absolute !important;
-  }
-  
-  .player1Card{
-    left: 32%;
-    top: 25%;
-  }
-  
-  .player2Card{
-    right: 32%;
-    top: 25%;
-  }
-  
-.tooltip {
-    display: none;
-}
-
-.disableCard {
-    pointer-events: none;
-}
-
-.LakersCardHome {
-    background-color: $lakers-pri; 
-    border: 5px double $lakers-sec;
-    color: $lakers-sec;
-}
-.HeatCardHome {
-    background-color: $heat-pri;
-    border: 5px double $heat-sec;
-    color: $heat-sec;
-}
-.RaptorsCardHome {
-    background-color: $raptors-pri;
-    border: 5px double $raptors-sec;
-    color: $raptors-sec;
-}
-.RocketsCardHome {
-    background-color: $rockets-pri;
-    border: 5px double $rockets-sec;
-    color: $rockets-sec;
-}
-
-.player1Card {
-    border-radius: 25px;
-    min-width: 125px;
-    max-width: 25%;
-    text-align: center;
-
-    &__img{
-        max-width: 60%;
-        min-width: 125px;
-        padding: 0;
-        margin: 0
-    }
-
-    &__Lakers2 {
-        background-color: $lakers-pri;
-        padding: 0.5rem;
-        max-width: 80%;
-        border-radius: 25px;
-        margin: auto !important;
-    }
-    &__Lakers1 {
-        background-color: $lakers-sec;
-        color: $lakers-pri !important;
-        padding: 0.5rem;
-        border-radius: 25px;
-        max-width: 80%;
-        margin: auto !important;
-    }
-    &__Heat2 {
-        background-color: $heat-pri;
-        padding: 0.5rem;
-        max-width: 80%;
-        border-radius: 25px;
-        margin: auto !important;
-    }
-    &__Heat1 {
-        background-color: $heat-sec;
-        padding: 0.5rem;
-        color: $heat-pri !important;
-        border-radius: 25px;
-        max-width: 80%;
-        margin: auto !important;
-    }
-    &__Raptors2 {
-        background-color: $raptors-pri;
-        padding: 0.5rem;
-        max-width: 80%;
-        border-radius: 25px;
-        margin: auto !important;
-        
-    }
-    &__Raptors1 {
-        background-color: $raptors-sec;
-        padding: 0.5rem;
-        color: $raptors-pri !important;
-        border-radius: 25px;
-        max-width: 80%;
-        margin: auto !important;
-    }
-    &__Rockets2 {
-        background-color: $rockets-pri;
-        padding: 0.5rem;
-        max-width: 80%;
-        border-radius: 25px;
-        margin: auto !important;
-        border-radius: 25px;
-    }
-    &__Rockets1 {
-        background-color: $rockets-sec;
-        padding: 0.5rem;
-        color: $rockets-pri !important;
-        border-radius: 25px;
-        max-width: 80%;
-        margin: auto !important;
-
-    }
-
-    &__stat:hover {
-        cursor: pointer;
-    }
-
-    &__stat:hover .tooltip {
-        display: block;
-    }
-
-    &__name:hover > .playertip {
-        display: block !important;
-    }
-
-}
-
-.player2Card {
-    border-radius: 25px;
-    min-width: 125px;
-    max-width: 25%;
-    text-align: center;
-
-    &__img{
-        max-width: 60%;
-        min-width: 125px;
-        padding: 0;
-        margin: 0
-    }
-
-    &__Lakers1 {
-        background-color: $lakers-pri;
-        padding: 0.5rem;
-        max-width: 80%;
-        margin: auto !important;
-        border-radius: 25px;
-    }
-    &__Lakers2 {
-        background-color: $lakers-sec;
-        color: $lakers-pri !important;
-        padding: 0.5rem;
-        border-radius: 25px;
-        max-width: 80%;
-        margin: auto !important;
-    }
-    &__Heat1 {
-        background-color: $heat-pri;
-        padding: 0.5rem;
-        max-width: 80%;
-        margin: auto !important;
-        border-radius: 25px;
-    }
-    &__Heat2 {
-        background-color: $heat-sec;
-        padding: 0.5rem;
-        color: $heat-pri !important;
-        border-radius: 25px;
-        max-width: 80%;
-        margin: auto !important;
-    }
-    &__Raptors1 {
-        background-color: $raptors-pri;
-        padding: 0.5rem;
-        max-width: 80%;
-        margin: auto !important;
-        border-radius: 25px;
-    }
-    &__Raptors2 {
-        background-color: $raptors-sec;
-        padding: 0.5rem;
-        color: $raptors-pri !important;
-        border-radius: 25px;
-        max-width: 80%;
-        margin: auto !important;
-    }
-    &__Rockets1 {
-        background-color: $rockets-pri;
-        padding: 0.5rem;
-        max-width: 80%;
-        margin: auto !important;
-        border-radius: 25px;
-    }
-    &__Rockets2 {
-        background-color: $rockets-sec;
-        padding: 0.5rem;
-        color: $rockets-pri !important;
-        border-radius: 25px;
-        max-width: 80%;
-        margin: auto !important;
-    }
-
-    &__stat:hover {
-        cursor: pointer;
-    }
-
-    &__stat:hover .tooltip {
-        display: block !important;
-    }
-
-    &__name:hover > .playertip {
-        display: block !important;
-    }
-
-}
-
-.playertip {
-    display: none !important;
-    background-color: blue !important;
-    color: white !important;
-    border-radius: 25px !important;
-    max-width: 60%;
-    margin: auto
-}
-
-.player1Card h4, .player2Card h4 {
-    margin: 0;
-}
-
-h4:active{
-    background-color: white;
-    border-radius: 25px;
-    color: black;
-}
-
-.playCard1:hover, .playCard2:hover, .stat:hover{
-    cursor: pointer
-}
