--- client/src/components/header.css
@@ -1,12 +1,0 @@
-#heading {
-    color: rgb(47, 39, 39);
-    position: absolute;
-    bottom: 5%;
-    left: 40%;
-    font-size: 2.9em;
-    display: flex;
-    justify-content: center;
-    margin: 0;
-    padding: 0;
-    text-transform: uppercase;
-}
