+++ client/src/pages/home.css
@@ -1,8 +1,9 @@
 .home {
     background-image: url("../images/bCourt.png");
+    /* background-image: url("../images/CourtBG.svg"); */
     height: 100%;
     width: 100%;
-    background-position: center;
+    /* background-position: center; */
     background-repeat: no-repeat;
     background-size: 100% 100%;
 }
